requirements.txt
curl_cffi>=0.5.0
gspread>=5.0.0
google-auth>=2.0.0
schedule>=1.2.0
pyperclip>=1.8.2




